preliminary unix support
